refactor(cuda): add reusable cuda_formula lib
